Consolidate shared CLI worker boot logic
Keep worker boot contracts, transfer policies, and spawned-child wiring in one documented module shared by the v1 and v2 workers. Inline the one-use spawn callbacks and explain mount state, child control endpoint ownership, and the main-thread lifecycle service.
