+++ packages/playground/cli/src/blueprints-v2/blueprints-v2-handler.ts
@@ -37,10 +37,7 @@ import {
 } from '../run-cli';
 import type { CLIOutput } from '../cli-output';
 import { cliExtensionArgsToExtensionsArray } from '../php-extensions';
-import {
-	workerBootApiTransferPolicy,
-	type WorkerConfig,
-} from '../worker-boot-config';
+import { workerBootApiTransferPolicy, type WorkerConfig } from '../worker-boot';
 
 /**
  * Boots Playground CLI workers using the native TypeScript Blueprint v2
