+++ packages/playground/cli/src/worker-boot.ts
@@ -1,10 +1,26 @@
+/**
+ * Shared boot protocol for Playground CLI workers.
+ *
+ * Both Blueprint worker implementations use this module to exchange boot
+ * configuration, transfer the MessagePorts a worker needs, and finish wiring a
+ * child created for proc_open() or system(). Keeping those contracts and the
+ * boot operation together prevents the v1 and v2 workers from implementing
+ * subtly different process behavior.
+ *
+ * Once the CLI uses the Blueprint v2 worker for both Blueprint formats, this
+ * shared code can move into that single worker implementation.
+ */
 import type { PHPExtension } from '@php-wasm/node';
 import {
+	consumeAPI,
 	defineAPITransferPolicy,
+	releaseApiProxy,
 	type AllPHPVersion,
 	type PathAlias,
+	type RemoteAPI,
 } from '@php-wasm/universal';
 import type { Mount } from '@php-wasm/cli-util';
+import { logger } from '@php-wasm/logger';
 import type { MessagePort } from 'worker_threads';
 
 export const CHILD_WORKER_CONTROL_READY = 'child-worker-control-ready';
@@ -144,3 +160,92 @@ export const workerBootApiTransferPolicy =
 			},
 		},
 	});
+
+type SpawnedChildWorkerApi = WorkerBootApi & {
+	useFileLockManager(port: MessagePort): Promise<void>;
+};
+
+const spawnedChildWorkerTransferPolicy =
+	defineAPITransferPolicy<SpawnedChildWorkerApi>({
+		bootRequestHandler: workerBootApiTransferPolicy.bootRequestHandler,
+	});
+
+/**
+ * Ask the main thread for a child worker and finish wiring it before exposing
+ * the child PHP proxy to the sandboxed spawn handler.
+ *
+ * The service returns all transferable ports in one response. Keeping the
+ * wiring here means the v1 and v2 workers only deal with a ready child PHP API,
+ * while the synchronous file-lock connection still goes directly from the
+ * child to the main thread.
+ */
+export async function bootSpawnedChildWorker<
+	WorkerApi extends SpawnedChildWorkerApi,
+>(
+	childWorkerService: RemoteAPI<ChildWorkerService>,
+	platformConfig: WorkerPlatformConfig
+) {
+	const { childId, phpPort, fileLockManagerPort, workerConfig } =
+		await childWorkerService.createChildWorker();
+	const runtimeExited = childWorkerService.waitForChildExit(childId);
+	let childApi: RemoteAPI<SpawnedChildWorkerApi> | undefined;
+
+	try {
+		const activeChildApi = consumeAPI<SpawnedChildWorkerApi>(
+			phpPort,
+			undefined,
+			spawnedChildWorkerTransferPolicy,
+			{ endpointTerminated: runtimeExited }
+		);
+		childApi = activeChildApi;
+		await activeChildApi.useFileLockManager(fileLockManagerPort);
+		await activeChildApi.bootRequestHandler(platformConfig, workerConfig);
+		// A child created after WordPress installation must not execute PHP until
+		// the service has applied the mounts it recorded for future workers.
+		await childWorkerService.waitForChildReady(childId);
+
+		return {
+			php: activeChildApi as unknown as RemoteAPI<WorkerApi>,
+			reap() {
+				releaseChildApiProxyQuietly(childApi);
+				// Reaping is deliberately fire-and-forget because the spawn-handler
+				// contract is synchronous. The service also has an exit backstop and
+				// disposeCLI() awaits any child still being terminated.
+				childWorkerService
+					.disposeChildWorker(childId)
+					.catch(function reportReapFailure(error) {
+						logger.error(
+							`Failed to reap child worker ${childId}: ${String(error)}`
+						);
+					});
+			},
+		};
+	} catch (error) {
+		releaseChildApiProxyQuietly(childApi);
+		try {
+			await childWorkerService.disposeChildWorker(childId);
+		} catch (disposeError) {
+			// Preserve the child boot failure; disposal is best-effort here and
+			// the service's exit backstop owns any remaining cleanup.
+			logger.error(
+				`Failed to dispose child worker ${childId} after a boot error: ` +
+					String(disposeError)
+			);
+		}
+		throw error;
+	}
+}
+
+/** Release a child proxy even when its worker already closed the endpoint. */
+function releaseChildApiProxyQuietly(
+	childApi: RemoteAPI<SpawnedChildWorkerApi> | undefined
+): void {
+	if (!childApi) {
+		return;
+	}
+	try {
+		childApi[releaseApiProxy]();
+	} catch {
+		// The remote endpoint may already have released the proxy on exit.
+	}
+}
